make the echo idl match the win32 IDL again (This used to be commit 4b901bd717a48daae24d8f7738169ac3cc5ba7fa)

+++ source4/librpc/ndr/ndr_echo.h
@@ -115,7 +115,7 @@ union echo_Info {
 
 struct TestCall2 {
 	struct {
-		uint32 level;
+		uint16 level;
 	} in;
 
 	struct {
